+++ sdk/serialization/azure-json/src/main/java/com/azure/json/implementation/jackson/core/json/ReaderBasedJsonParser.java
@@ -1244,15 +1244,7 @@ public class ReaderBasedJsonParser extends ParserBase {
                 }
             }
             ch = _inputBuffer[_inputPtr++];
-            if (ch == 'N') {
-                String match = negative ? "-INF" : "+INF";
-                _matchToken(match, 3);
-                if ((_features & FEAT_MASK_NON_NUM_NUMBERS) != 0) {
-                    return resetAsNaN(match, negative ? Double.NEGATIVE_INFINITY : Double.POSITIVE_INFINITY);
-                }
-                _reportError(
-                    "Non-standard token '" + match + "': enable JsonParser.Feature.ALLOW_NON_NUMERIC_NUMBERS to allow");
-            } else if (ch == 'n') {
+            if (ch == 'n') {
                 String match = negative ? "-Infinity" : "+Infinity";
                 _matchToken(match, 3);
                 if ((_features & FEAT_MASK_NON_NUM_NUMBERS) != 0) {