+++ sdk/serialization/azure-json/src/test/java/com/azure/json/contract/JsonReaderContractTests.java
@@ -10,6 +10,8 @@ import com.azure.json.JsonWriter;
 import com.azure.json.ReadValueCallback;
 import org.junit.jupiter.api.Named;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -445,9 +447,9 @@ public abstract class JsonReaderContractTests {
 
     private static Stream<Arguments> readUntypedSimpleSupplier() {
         return Stream.of(Arguments.of("null", 1, null), Arguments.of("true", 1, true), Arguments.of("false", 1, false),
-            Arguments.of("3.14", 1, 3.14), Arguments.of("NaN", 1, String.valueOf(Double.NaN)),
-            Arguments.of("-Infinity", 1, String.valueOf(Double.NEGATIVE_INFINITY)),
-            Arguments.of("Infinity", 1, String.valueOf(Double.POSITIVE_INFINITY)), Arguments.of("42", 1, 42),
+            Arguments.of("3.14", 1, 3.14D), Arguments.of("NaN", 1, Double.NaN),
+            Arguments.of("-Infinity", 1, Double.NEGATIVE_INFINITY),
+            Arguments.of("Infinity", 1, Double.POSITIVE_INFINITY), Arguments.of("42", 1, 42),
             Arguments.of("420000000000", 1, 420000000000L), Arguments.of("\"hello\"", 1, "hello"));
     }
 
@@ -745,6 +747,22 @@ public abstract class JsonReaderContractTests {
             Arguments.of("{\"boolean\":true,\"int\":42,\"decimal\":42.0,\"string\":\"hello\"}", reader, map));
     }
 
+    @Execution(ExecutionMode.SAME_THREAD)
+    @ParameterizedTest
+    @MethodSource("readUntypedExponentNumbersSupplier")
+    public void readUntypedExponentNumbers(String numberString, Number expected) throws IOException {
+        readAndValidate(numberString, JsonReader::readUntyped, actual -> assertEquals(expected, actual));
+    }
+
+    private static Stream<Arguments> readUntypedExponentNumbersSupplier() {
+        return Stream.of(Arguments.of("1e-1", 0.1D), Arguments.of("1E-1", 0.1D), Arguments.of("1e+1", 10D),
+            Arguments.of("1E+1", 10D), Arguments.of("1e-01", 0.1D), Arguments.of("1E-01", 0.1D),
+            Arguments.of("1e+01", 10D), Arguments.of("1E+01", 10D), Arguments.of("1e0", 1D), Arguments.of("1E0", 1D),
+
+            Arguments.of("Infinity", Double.POSITIVE_INFINITY), Arguments.of("+Infinity", Double.POSITIVE_INFINITY),
+            Arguments.of("-Infinity", Double.NEGATIVE_INFINITY), Arguments.of("NaN", Double.NaN));
+    }
+
     private static void assertJsonReaderStructInitialization(JsonReader reader, JsonToken expectedInitialToken)
         throws IOException {
         assertNull(reader.currentToken());
@@ -772,9 +790,8 @@ public abstract class JsonReaderContractTests {
             return test;
         }
 
-        public TestData setTest(String test) {
+        public void setTest(String test) {
             this.test = test;
-            return this;
         }
 
         @Override