+++ sdk/serialization/azure-json/src/test/java/com/azure/json/implementation/DefaultJsonReaderContractTests.java
@@ -6,7 +6,6 @@ package com.azure.json.implementation;
 import com.azure.json.JsonOptions;
 import com.azure.json.JsonReader;
 import com.azure.json.contract.JsonReaderContractTests;
-import com.azure.json.implementation.jackson.core.JsonParseException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
@@ -20,8 +19,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
  */
 public class DefaultJsonReaderContractTests extends JsonReaderContractTests {
     private JsonReader reader;
-    String jsonWithComments = "{    // single line comment\n" + "    \"single-line\": \"comment\",\n" + "    /*\n"
-        + "    multi-line comment\n" + "    */\n" + "    \"multi-line\": \"comment\"}";
+    String jsonWithComments = "{// single line comment\n\"single-line\":\"comment\","
+        + "\n/*\nmulti-line comment\n*/\n\"multi-line\":\"comment\"}";
 
     @Override
     public JsonReader getJsonReader(String json) throws IOException {
@@ -38,7 +37,6 @@ public class DefaultJsonReaderContractTests extends JsonReaderContractTests {
 
     @Test
     public void readJsonc() throws IOException {
-
         try (JsonReader jsonReader
             = DefaultJsonReader.fromString(jsonWithComments, new JsonOptions().setJsoncSupported(true))) {
             jsonReader.nextToken();
@@ -48,8 +46,8 @@ public class DefaultJsonReaderContractTests extends JsonReaderContractTests {
     }
 
     @Test
-    public void readJsoncFails() throws IOException {
-        assertThrows(JsonParseException.class, () -> {
+    public void readJsoncFails() {
+        assertThrows(IOException.class, () -> {
             try (JsonReader jsonReader = getJsonReader(jsonWithComments)) {
                 jsonReader.nextToken();
                 String outputJson = jsonReader.readChildren();